Move update to mono source set
