+++ app/shared/src/androidMain/kotlin/tools/update/AndroidUpdateInstaller.kt
@@ -1,4 +1,4 @@
-package me.him188.ani.app.update
+package me.him188.ani.app.tools.update
 
 import android.content.Context
 import android.content.Intent
