+++ app/shared/src/commonMain/kotlin/ui/update/FileDownloaderPresentation.kt
@@ -1,11 +1,11 @@
-package me.him188.ani.app.update.ui
+package me.him188.ani.app.ui.update
 
 import androidx.compose.runtime.Stable
 import androidx.compose.runtime.State
 import androidx.compose.runtime.getValue
 import kotlinx.coroutines.CoroutineScope
+import me.him188.ani.app.tools.update.FileDownloader
 import me.him188.ani.app.ui.foundation.HasBackgroundScope
-import me.him188.ani.app.update.FileDownloader
 
 /**
  * 将 [FileDownloader] 状态收集为 Compose [State]
